Parcel now works
I found a solution, it is completely different from the task instructions.

How did I do it?
1. I installed "parcel-bundler" as dependency.
2. in the package.json file in "scripts" I created a new "dev" script: "parcel src/index.htm".
3. In the terminal at the location of my project, I typed "npm run dev".

with the previous steps I did create the folder "dist" with the .js, .css, .map files and an extra .js.map file. also I added the sass in dev dependencies inside the package.json file and when I open the locolhost:1234 the web page does not give me an error.

I still have the doubt why it didn't work with the method described in the lesson?

Now with this new method, instead of using "parcel[src/index.html]" in the terminal, now I will use "npm run dev"?