Integrating the Login View
LoginView is imported into MainView at the top of the code. This is because LoginView will need to get the user details from the MainView. If LoginView is not imported here, there would be no way of passing the user details to it.

The user property is initialized to null in the state (default is logged out). When the app is first run or when a user has logged out, there is no user that is logged in, hence setting the user to null.

A method, onLoggedIn, will be passed as a prop with the same name to LoginView (note that this just happens to be the same name—it’s not a constraint). This method will update the user state of the MainView component and will be called when the user has successfully logged in.

The LoginView is rendered as long as there's no user in the state. This is because the user property will be null in the MainView state.